Finalize MVP: all screens wired, compile-audit fixes, capabilities
Compile-audit fixes:
- jobs::list_filtered: switch to positional params via params_from_iter
  (avoids rusqlite named-param lifetime edge cases).
- commands/resumes::tailor_resume_for_job: hold owned Strings for paths
  instead of &Cow<str> borrows into temporaries; clear lifetime story.
- ingestion/runner: persist_one now returns PersistOutcome::Prepared(id)
  so ingest_one returns the real job id; parent folder string bound to
  an owned value before passing by reference.
- ingestion_runs::finish: cast u32 totals to i64 for rusqlite's ToSql.
- lib.rs setup(): clone app.handle() explicitly and pass by reference
  (compat with either owned or borrowed handle APIs).
- src/vite-env.d.ts added so CSS-module and @/ alias imports typecheck.
- src-tauri/capabilities/default.json added so Tauri 2 starts with the
  core + opener + shell + dialog plugin permissions available.

Screens completed (all wired to real IPC):
- JobDetail: score breakdown (8-term bar chart), sponsorship panel,
  file links, account list for the company, resume variants for the
  job, full JD text, status dropdown, tailor-resume action.
- Accounts: list + per-card save-to-keychain, reveal, clear, delete,
  open-portal (marks last_used_at), add-account dialog that ships the
  initial password straight to the keychain on create.
- Companies: name + folder path + job count + open-folder action;
  links into Accounts for portal management.
- RunHistory: last 50 runs with per-source fetched/filtered/prepared/
  errors table.
- ResumeManager: JSON editor for the canonical profile, apply & save,
  lightweight preview summary; save persists to the configured path
  (or the default ~/Documents/JobSearchTool/master_resume.json).

docs/PACKAGING.md notes the `tauri icon` prerequisite for `tauri build`.

+++ src-tauri/src/db/repo/ingestion_runs.rs
@@ -37,7 +37,14 @@ pub fn finish(
             total_fetched = ?2, total_filtered = ?3, total_prepared = ?4,
             notes = ?5
          WHERE id = ?6",
-        params![json, totals.0, totals.1, totals.2, notes, id],
+        params![
+            json,
+            totals.0 as i64,
+            totals.1 as i64,
+            totals.2 as i64,
+            notes,
+            id,
+        ],
     )?;
     Ok(())
 }
